Remove preload.js in plugin uses and use front plugin injection
